+++ src/arch/netlrts/machine.C
@@ -1648,10 +1648,10 @@ int DeliverOutgoingMessage(OutgoingMsg ogm)
 /**
  * Set up an OutgoingMsg structure for this message.
  */
-static OutgoingMsg PrepareOutgoing(int pe,int size,int freemode,char *data) {
+static OutgoingMsg PrepareOutgoing(int pe,size_t size,int freemode,char *data) {
   OutgoingMsg ogm;
   MallocOutgoingMsg(ogm);
-  MACHSTATE2(2,"Preparing outgoing message for pe %d, size %d",pe,size);
+  MACHSTATE2(2,"Preparing outgoing message for pe %d, size %zu",pe,size);
   ogm->size = size;
   ogm->data = data;
   ogm->src = CmiMyPeGlobal();
@@ -1674,8 +1674,8 @@ static OutgoingMsg PrepareOutgoing(int pe,int size,int freemode,char *data) {
  *
  *****************************************************************************/
 
-//CmiCommHandle CmiGeneralSend(int pe, int size, int freemode, char *data)
-CmiCommHandle LrtsSendFunc(int destNode, int pe, int size, char *data, int freemode)
+//CmiCommHandle CmiGeneralSend(int pe, size_t size, int freemode, char *data)
+CmiCommHandle LrtsSendFunc(int destNode, int pe, size_t size, char *data, int freemode)
 {
   int sendonnetwork;
   OutgoingMsg ogm;
@@ -1704,8 +1704,8 @@ CmiCommHandle LrtsSendFunc(int destNode, int pe, int size, char *data, int freem
  * NET version List-Cast and Multicast Code
  *
  ****************************************************************************/
-                                                                                
-void LrtsSyncListSendFn(int npes, const int *pes, int len, char *msg)
+
+void LrtsSyncListSendFn(int npes, const int *pes, size_t len, char *msg)
 {
   int i;
   for(i=0;i<npes;i++) {
@@ -1713,8 +1713,8 @@ void LrtsSyncListSendFn(int npes, const int *pes, int len, char *msg)
     CmiSyncSendAndFree(pes[i], len, msg);
   }
 }
-                                                                                
-CmiCommHandle LrtsAsyncListSendFn(int npes, const int *pes, int len, char *msg)
+
+CmiCommHandle LrtsAsyncListSendFn(int npes, const int *pes, size_t len, char *msg)
 {
   CmiError("ListSend not implemented.");
   return (CmiCommHandle) 0;
@@ -1725,7 +1725,7 @@ CmiCommHandle LrtsAsyncListSendFn(int npes, const int *pes, int len, char *msg)
   returns is not changed, we can use memory reference trick to avoid 
   memory copying here
 */
-void LrtsFreeListSendFn(int npes, const int *pes, int len, char *msg)
+void LrtsFreeListSendFn(int npes, const int *pes, size_t len, char *msg)
 {
   int i;
   for(i=0;i<npes;i++) {
@@ -2258,7 +2258,7 @@ void LrtsInit(int *argc, char ***argv, int *numNodes, int *myNodeID)
 }
 
 
-void LrtsPrepareEnvelope(char *msg, int size)
+void LrtsPrepareEnvelope(char *msg, size_t size)
 {
   CMI_MSG_SIZE(msg) = size;
 }
